backend/src/requirements.txt
pymongo==3.11.0
Flask-PyMongo==2.3.0
python-dotenv==0.19.0
Flask-Cors==3.0.9
shortuuid==1.0.1
dnspython==1.16.0